+++ src/main/java/server/StorageInventory.java
@@ -19,18 +19,13 @@
 */
 package server;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
-
 import client.Client;
 import client.inventory.Equip;
 import client.inventory.Item;
 import config.YamlConfig;
 import constants.inventory.ItemConstants;
+
+import java.util.*;
 
 /**
  * @author RonanLana
@@ -38,11 +33,11 @@ import constants.inventory.ItemConstants;
 public class StorageInventory {
     private final Client c;
     private Map<Short, Item> inventory = new LinkedHashMap<>();
-    private final short slotLimit;
+    private final byte slotLimit;
 
     public StorageInventory(Client c, List<Item> toSort) {
         this.inventory = new LinkedHashMap<>();
-        this.slotLimit = (short) toSort.size();
+        this.slotLimit = (byte) toSort.size();
         this.c = c;
 
         for (Item item : toSort) {
@@ -50,7 +45,7 @@ public class StorageInventory {
         }
     }
 
-    private short getSlotLimit() {
+    private byte getSlotLimit() {
         return slotLimit;
     }
